Refactor data layer into its own manager
This change prepares the app for adding async storage support and
for adding categories. Adding async storage is the point at which
it will get harder to change the data layer, so I made some changes
to respect categories (not yet implemented) with an eye toward
having draggable items and categories in the future.